move ordinal to separate module
